Creato effetto lerp per variare la vita del boss(sfera al centro della stanza del boss, premendo Q abbassi la vita di 5, il boss ha 100 di vita, aggiunto un effetto nuovo al menu di inizio, premi play per vederlo, ciao

+++ The Dream of a Ball/Assets/Scripts/Rotator.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class Rotator : MonoBehaviour
 {
-    public int Y;
-    public int Z;
-    public int X;
+    public float Y;
+    public float Z;
+    public float X;
     // Use this for initialization
     void Start()
     {
